+++ app/src/main/java/cca/capitalcityaquatics/aquariuminfo/data/calculators/FlowRateData.kt
@@ -1,13 +1,14 @@
 package cca.capitalcityaquatics.aquariuminfo.data.calculators
 
+import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 
 data class FlowRateData (
+	@DrawableRes val image: Int,
 	@StringRes val subtitle: Int,
-	@StringRes val subtitle2: Int,
 	@StringRes val formulaText: Int,
-	val conversionLowFreshwater: Double,
-	val conversionHighFreshwater: Double,
-	val conversionLowReef: Double,
-	val conversionHighReef: Double,
+	@StringRes val icon: Int,
+	@StringRes val header: Int,
+	val conversionLow: Double,
+	val conversionHigh: Double,
 )