+++ app/src/main/java/cca/capitalcityaquatics/aquariuminfo/data/calculators/FlowRateDataSource.kt
@@ -1,14 +1,27 @@
 package cca.capitalcityaquatics.aquariuminfo.data.calculators
 
 import cca.capitalcityaquatics.aquariuminfo.R
+import cca.capitalcityaquatics.aquariuminfo.navigation.FishCompatibilityFreshwater
+import cca.capitalcityaquatics.aquariuminfo.navigation.FishCompatibilityMarine
 
-val flowRateDataSource =
+val flowRateDataSourceFreshwater =
 	FlowRateData(
-		subtitle = R.string.text_subtitle_alk,
-		subtitle2 = R.string.text_subtitle_alk_1,
-		formulaText = R.string.text_formula_alk,
-		conversionLowFreshwater = 3.0,
-		conversionHighFreshwater = 5.0,
-		conversionLowReef = 5.0,
-		conversionHighReef = 10.00,
+		image = R.drawable.freshwater_shark,
+		subtitle = R.string.flow_rate_subtitle,
+		formulaText = FishCompatibilityFreshwater.title,
+		icon = FishCompatibilityFreshwater.icon,
+		header = R.string.flow_rate_freshwater,
+		conversionLow = 3.0,
+		conversionHigh = 5.0,
+	)
+
+val flowRateDataSourceMarine =
+	FlowRateData(
+		image = R.drawable.clownfish,
+		subtitle = R.string.flow_rate_subtitle,
+		formulaText = FishCompatibilityMarine.title,
+		icon = FishCompatibilityMarine.icon,
+		header = R.string.flow_rate_marine,
+		conversionLow = 5.0,
+		conversionHigh = 10.00,
 	)